requirements.txt
flask==3.0.2
flask-socketio==5.3.6
python-dotenv==1.0.1
PyPDF2==3.0.1
openai==1.75.0
pandas==2.2.1
gevent==23.9.1
gevent-websocket==0.10.1
markdown==3.5.2
sentence-transformers==2.5.1